Resulto registro sin image profile

+++ api/src/controllers/patients.controllers.js
@@ -37,7 +37,16 @@ function concat_json(json, json_empty) {
 }
 
 const createPatient = async (req, res) => {
-  let result = null;
+  let result;
+  try {
+    if (req.file != undefined) {
+      result = await cloudinary.v2.uploader.upload(req.file.path);
+      result = result.url;
+    }
+  } catch (error) {
+    console.log("Error con cloudinary:", error);
+  }
+
   const {
     dni,
     name,
@@ -48,9 +57,6 @@ const createPatient = async (req, res) => {
     num_member,
     healthInsuranceId,
   } = req.body;
-  if (req.file.path) {
-    result = await cloudinary.v2.uploader.upload(req.file.path);
-  }
   const rol = "Patient";
   if (
     dni &&
@@ -69,7 +75,7 @@ const createPatient = async (req, res) => {
           name,
           lastname,
           address,
-          imageProfile: result.url,
+          imageProfile: result,
           email,
           password: encryptPassword(password),
           rol,
@@ -87,7 +93,13 @@ const createPatient = async (req, res) => {
           ],
         }
       );
-      await fs.unlink(req.file.path); // Elimina la imagen guardada en api/src/public/uploads
+      try {
+        if (req.file != undefined) {
+          await fs.unlink(req.file.path); // Elimina la imagen guardada en api/src/public/uploads
+        }
+      } catch (error) {
+        console.log("Error eliminando la imagen guardada", error);
+      }
       let newPatient = await Patient.create(
         {
           num_member,
